Get-RuntimeAcquisitionInfo -- improved summary writing

+++ .cicd_pipeline_helpers/ps_modules/runtime_env_info_display/Public/Get-RuntimeAcquisitionInfo.ps1
@@ -34,12 +34,29 @@ Function Get-RuntimeAcquisitionInfo {
         If ([System.Environment]::GetEnvironmentVariable('TF_BUILD') -eq 'True') {
             # Credit:  https://stackoverflow.com/a/77511799
             $tempFilePath = [System.IO.Path]::GetFullPath([System.IO.Path]::Combine($ADOAgentTempDirectoryPath, "temp.md"))
-            Write-Host 'Yes, unit test, we reached this'
+            # Write-Host "We will be writing to:  ${tempFilePath}" # DEBUG LINE ONLY
+            If ([System.Environment]::GetEnvironmentVariable('IS_JUST_PESTER_TEST') -eq 'True') {
+                Write-Host 'Yes, unit test, we reached this'
+            }
+            # Add-Content `
+            #     -Path "$tempFilePath" `
+            #     -Value "# Acquisition time summary"
+            # Add-Content `
+            #     -Path "$tempFilePath" `
+            #     -Value ""
             Add-Content `
                 -Path "$tempFilePath" `
                 -Value "Acquisition time: ${acquisitionSec}s"
-            Write-Host "##vso[task.uploadsummary]${tempFilePath}" # ADO-specific
-            Remove-Item -Path "$tempFilePath" -Force -ErrorAction 'SilentlyContinue'
+            # $pathContents = Get-Content -Path "$tempFilePath" # DEBUG LINE ONLY
+            # Write-Host 'Dumping summary contents' # DEBUG LINE ONLY
+            # Write-Host "$pathContents" # DEBUG LINE ONLY
+            # Write-Host 'Dumped summary contents' # DEBUG LINE ONLY
+            Write-Host "##vso[task.addattachment type=Distributedtask.Core.Summary;name=AgentAcquisitionTime;]${tempFilePath}" # ADO-specific
+            # Write-Host 'Allegedly uploaded summary' # DEBUG LINE ONLY
+            If ([System.Environment]::GetEnvironmentVariable('IS_JUST_PESTER_TEST') -eq 'True') {
+                Remove-Item -Path "$tempFilePath" -Force -ErrorAction 'SilentlyContinue'
+                # Write-Host 'Removed summary file' # DEBUG LINE ONLY
+            }
         }
     } # end PROCESS
     End {} # end END
